Remove reset_area method
The `reset_area` method has been created because it was used by both the
BarComponent, for clearing it, and the event loop for handling expose
events. However it is not necessary to reset the area back to the
background color after an expose event, but it is necessary for clearing
a component. Because of this differentiation the `reset_area` method has
been removed and the relevant content of it has been moved to the place
where it has been called from.