Add font rendering
This is the first version of the bar that supports font rendering on a
basic level. It is possible to set the text, the desired font with the
size (pango style) and the alignment of that text inside the component.

If the background is smaller than the width of the text, the background
stays at the left side of the component and the width is extended. In
future revisions it might be nice to offset background according to font
alignment.

A remaining issue with backgrounds smaller than the text is that the
background behind the text if no other background is set is just a
garbled mess. This is probably because the pixmap does not support
transparency and will be automatically fixed when transparent pngs also
are supported.