Optimize locking and X resources
There was a locking bug in the code, which lead to each component
waiting until after the `sleep` before releasing the lock. This has been
changed and a scope has been added. The lock is now only active in the
immediate scope.

XCB also had a bit of resource leakage with pixmaps being created, but
never freed. Now all pixmaps are freed after the pictures are composited
and no X resources are leaked anymore.