Fix default font centering bug
There was a bug where the fallback font `FontDescription::new()` would
not be centered correctly because the `size` of the font is `0`. This
has been fixed by changing the font centering code to use
`layout.get_pixel_size()', which works with the empty description. It's
also a lot cleaner than the previous solution.

Because the `text_size` method is only used in `bar.rs` to get the
width, it has been changed to `get_width` and does not return height
anymore.