Replace unwraps with error handling
Every unwrap that is not intended has been removed and replaced by
handling the error. Most errors now just return an `XError` with an id.
This could be improved by adding more explicit error messages.

The general idea inside when drawing a component is to just interrupt
the redraw and try again next time. This might lead to issues with
components not properly being cleaned up or redrawn, but it's not
something that should ever happen (errors with the X server).

To make these changes possible and easier to implement, a lot of code
has been moved into different modules, this made it easier to just
return error values and then handle it with the caller. It also allowed
cleaning up part of the code a bit.