Add max/fixed width restrictions
Width is now completely separated from the background. It allows to set
min width, max width and a fixed width.

This allows the user to use an image that's bigger than the text width
without having a massive component.

The `ignore_text` and `ignore_background` flags also have been
introduced. This makes it easy to just scale a background to the width
of the text.