Change rendering pipeline
The basic rendering pipeline has been adjusted to fix flickering and
rendering issues. All components should now be drawn correctly without
weird flickering.

The issues were fixed by rendering to an intermediate picture first,
which is then copied to the background. Clearing areas of the bar is
only done when the components shrunk, otherwise the current data is just
overwritten.

This fixes #6 and fixes #7.