Improve performance
See #12.

1) Create custom image format

Images are now created by the user with a reference to the bar. This
allows to make images pictures that are stored in X. This improves
performance by preventing excessive re-encoding.

2) Don't render components without width change

If the width of a component did not change, the other components that
might usually be affected, are not redrawn. This is because those
components did not move. This should improve performance by reducing
redrawing significantly.

2) Create custom text format

The image format has been reworked a bit and similar to it, a new text
format has been created. These two types now allow caching images and
text using X, so stuff does not have to be re-rendered every X seconds.
These are created by the user and use reference-counting to make sure
the destructor is run once when all instances are gone.