Streamline user-facing API
There have been some minor tweaks to the way this library is used to
make it nicer to use.

This are some of the changes:
 * The foreground message now doesn't return an option anymore
 * The foreground itself now is empty by default
 * From<Text> is now implemented for Foreground
 * From<Image> is implemented for Background
 * Foreground and Background don't allow refs anymore (no hidden clone)

All these changes are intermediate steps for making the API more usable
for the client. Before these changes are accepted as permanent
solutions, more testing is required.