Fix component size bug
There have been two bugs in the renderer that are related to component
size.

One of them was an issue where when width or height are 0, the
update method attempted to create an empty pixmap, this does not work.
To fix this the method now returns early instead of updating

The other issue was related to the updating without foreground or
background. This did not create an empty pixmap when only a size
existed. To make it possible to have a component with width but without
foreground or background, the component now also updates when the
geometry changes.

This fixes #20.