Rework documentation
Documentation has been reworked so that every method now includes
examples instead of having one massive block of examples at the top of
every module. This should make it obvious and simple how methods should
be used.

Where convenient the bigger documentation blocks for the modules have
been kept in place to document the combination of the different methods.

This fixes #19.