Add time example
This change added a time example which shows how to display a simple
clock that updates on a timer. This is an implementation for the timer
example mentioned in #23.

Besides the addition of the new example the other examples have been
reworked. Documentation has been expanded and improved and some
unnecessary, confusing code was stripped out to keep them focused and
easy to understand.