Fix width calculation bug
When calculating a fixed width it would just return the fixed width
without regards of how big the bar actually is. This lead to an error
because XCB was not able to manage a pixmap with that size.

To fix this the max width has been limited to the width of the bar.

This fixes #25.