Refactor directory structure
Refactored the directory structure to make it a little more intuitive to
what does what. Generally there's now a place for everything component
related, which is most of the code, and a place for all the utility
functionality that is used in multiple places.

This directory structure might make it a little easier to start creating
errors. So #26 might be a good next issue to tackle.