Add layers to component::Background
The Background is now able to take multiple images instead of just a
single one. This allows the user to layer images above each other.

The alignment of the background can now be controlled for every image
separately and there is no option to align the background anymore, as a
result of that this is an API-breaking commit.

This fixes #22.